translate/requirements.txt
psycopg2-binary
python-dotenv
rdflib
requests
